Rest: remove Xml input/output type.
This type caused an extra parse/pretty print sequence to be performed, since it was output using its pickler, which first transformed the contained (xml) string to an xml DOM, and then pretty printed that. Now, we directly return the raw string/bytestring.

This fixes a bug in the TR. The XML output by the TR is parseable as HTML (renders <div></div> instead of <div/>), but the extra roundtrip through HXT broke this.