Upgrade hxt to latest version.
The escaping function `escapeXmlDoc` was removed, as apparently this functionality is now under the hood somewhere. However, `xshow` doesn't include it. Luckily, we can switch to `writeDocumentToString`.

Also, the internal structure of the `PU` type changed a bit, but nothing substantial, just currying of the functions.

Finally, the picklers in hxt now have proper error messages, so we can get rid of our own branch. For this, we have to call `unpickleDoc'` (with a prime) instead of `unpickleDoc`.