Rest-core: more serialization options for Reason.
If a handler cannot throw any errors (NoE), the reasonE function said that we couldn't output in any format. But we can output Reason in either Xml or Json, so we return that instead.

Before this change, you could not use a multi-PUT on endpoints that didn't throw any errors, e.g. /site/uri/<uri>/permission/id.