Resource: identify resources by id types.
These are generated by a schema that defines (part of) the routing for the resource. This allows us to reason about identities, pass them on to subhandlers, and in general, makes resource identification more first class. It also allows us to skip database queries to get the actual resource to pass to subhandlers. Additionally, it will make routing less ambiguous.