Rest: Handler takes an id now.
We no longer expose the actual returned type, but we do specify if it's a list of a single value.

Further, adapt all the dictionaries to remove the id from there.

Finally, remove the 'prepare' function. It was a hack to avoid too much work to pass a value to subhandlers, which is not going to happen anymore.