Rest-core: Remove 'id' from Handler.
Instead, 'list' and 'actions' get it as an argument. This way they can dispatch and change dictionaries based on it. The rest of the handlers get it as part of the sub-monad.

Also changed the 'By' identification to be more constrained. You now give an Ident again, giving us access to the Info for documentation, as well as a way to always get an 'id' (by passing undefined as the 'a' to the function 'a -> id').