Rest-core: better representation of no dictionaries.
Previously, you could have things like [NoO, JsonO] which makes no sense.

The dictionary modifiers like listO now return a Maybe, where Nothing indicates the case where none of the original dictionaries can be used with those List has. E.g. you can have RawXmlO as your output type, but we cannot wrap that in a list, since that only has Json and Xml instances.

This extends to mkListHandler and mkMultiPuthandler, which now also don't return a handler when it is not possible to generate the new dictionaries.