Rest-gen: Do not use .cabal templates but update Exposed-Modules in existing .cabal.
The downside of the templates is that they become out-of-date since bumper misses the templates. Changing bumper is not an option as the templates are not valid Cabal syntax and cannot be parsed.  Therefore, assume that a valid .cabal file exists and only update the Exposed-Modules upon client generation. An initial .cabal file is generated to get you started.

Note that now that the complete .cabal file is generated, the generation function needs to know what additional sources (i.e., modules that are not automatically generated but simply copied) should be included in the Exposed-Modules section.