Indicate empty dictionary list in types.
The type of the dictionary is now a type-level Maybe, with Nothing
indicating an empty list of dictionaries. This allows us to choose,
based on the context, if we want to treat the empty list as
polymorphic, or as monomorphic (e.g. ()). This means we no longer need
to `some*` combinators: we can just allow an empty dictionary to be
added to. However, in the driver code we treat empty dictionaries as
(), as before.

Fixes #31.