Rest-client: fix encoding issue with String inputs.
These were converted to lazy bytestrings using `fromString`, but that
function encodes as ASCII, not UTF-8, apparently.

I fixed this by exporting a `toLbs` function and using that. I did it
this way to avoid having to figure out how to best include an import,
since this is pretty deep in the code.